Moved game logic to other file
